feat(conversations): add JWT-only internal ticket route
PR 2 of #82564: an internal conversations ticket route verified by scoped
service JWTs (first consumer of the #82736 machinery), plus a counter that
tracks auth method on both ticket routes so the secret_api_token cutover
can watch legacy usage reach zero. The external route's handlers move
unchanged to a shared module; external behavior is identical.

Generated-By: PostHog Desktop
Task-Id: a4d3c877-c693-4947-9210-7ede72193444
