+++ posthog/jwt.py
@@ -24,6 +24,7 @@ class PosthogJwtAudience(Enum):
     WORKFLOWS_CANCEL_BATCH = "posthog:workflows:cancel_batch"
     INTEGRATION_SERVICE = "posthog:integration_service"
     TASKS_CREATE = "posthog:tasks:create"
+    CONVERSATIONS_TICKETS = "posthog:conversations:tickets"
 
 
 def signing_key_fingerprint(key: str) -> str:
