+++ products/conversations/backend/metrics.py
@@ -1,4 +1,13 @@
-from prometheus_client import Histogram
+from prometheus_client import Counter, Histogram
+
+# Cutover observability for #82564: the CDP worker's ticket actions move from
+# secret_api_token on the external route to scoped JWTs on the internal route.
+# The legacy worker path can be removed once auth_method="secret_api_token" stays at zero.
+TICKET_ACTION_AUTH_COUNTER = Counter(
+    "posthog_conversations_ticket_action_auth_total",
+    "Successful authentications on the ticket routes called by CDP workflow actions, by auth method",
+    labelnames=["auth_method", "http_method"],  # auth_method: secret_api_token | scoped_jwt
+)
 
 TICKET_SEARCH_DURATION_SECONDS = Histogram(
     "posthog_support_ticket_search_duration_seconds",
